Variable activity is uin32_t, and so is similar to that of precosat

+++ Solver/VarReplacer.cpp
@@ -101,8 +101,8 @@ const bool VarReplacer::performReplaceInternal()
         solver.setDecisionVar(var, false);
         solver.setDecisionVar(it->var(), true);
         
-        double& activity1 = solver.activity[var];
-        double& activity2 = solver.activity[it->var()];
+        uint32_t& activity1 = solver.activity[var];
+        uint32_t& activity2 = solver.activity[it->var()];
         if (wasDecisionVar && activity1 > activity2) {
             activity2 = activity1;
             solver.order_heap.update(it->var());
